Update planets.py
Added a color label to the earth class.
Created CelestialObject class, to develop all the calculations for the objects. Because before, we were doing the calculations in the class itself, but later we are going to repeated when adding more objects, in order to do not repeat the code, we created a class apart that's able to calculate this for us